backend/requirements.txt
Flask==3.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.1
Flask-JWT-Extended==4.7.1
scikit-learn>=1.5.0
numpy>=1.24.0,<2.0.0
pandas>=2.1.0
Pillow==11.1.0
reportlab==4.3.1
networkx==3.4.2
werkzeug==3.1.3
python-dotenv==1.0.1
joblib==1.4.2
bcrypt==4.2.1
kagglehub>=0.3.0
matplotlib>=3.8.0
tensorflow>=2.10.0
